server/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.3
pydantic==2.9.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
python-multipart==0.0.9
aiofiles==24.1.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
psycopg2-binary==2.9.9